[typescript] fix structure
