+++ typescript/.gitignore
@@ -1,2 +1,1 @@
 /build/
-/package-lock.json
