--- typescript/notebooks/.python-version
@@ -1,1 +1,0 @@
-3.7.5
