+++ typescript/package.json
@@ -1,23 +1,23 @@
 {
-    "name": "typescript.demo",
-    "version": "1.0.0",
-    "scripts": {
-        "build": "./node_modules/.bin/tsc --build tsconfig.json",
-        "watch": "./node_modules/.bin/tsc --build tsconfig.json --watch",
-        "startts": "./node_modules/.bin/ts-node index.ts",
-        "startjs": "node ./build/index.js",
-        "test": "./node_modules/.bin/mocha --require ts-node/register test/**/*.ts --color"
-    },
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "ts-node": "^8.6.2",
-        "typescript": "^3.7.5"
-    },
-    "devDependencies": {
-        "@types/jasmine": "^3.5.4",
-        "@types/node": "^13.7.1",
-        "chai": "^4.2.0",
-        "mocha": "^7.0.1"
-    }
+  "name": "typescript.demo",
+  "version": "1.0.0",
+  "scripts": {
+    "build": "./node_modules/.bin/tsc --build tsconfig.json",
+    "watch": "./node_modules/.bin/tsc --build tsconfig.json --watch",
+    "startts": "./node_modules/.bin/ts-node app/index.ts",
+    "startjs": "node ./build/app/index.js",
+    "test": "./node_modules/.bin/mocha --require ts-node/register app/test/**/*.ts --color"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "ts-node": "^8.6.2",
+    "typescript": "^3.7.5"
+  },
+  "devDependencies": {
+    "@types/jasmine": "^3.5.4",
+    "@types/node": "^13.7.1",
+    "chai": "^4.2.0",
+    "mocha": "^7.0.1"
+  }
 }
